feat: adicionados repositories no projeto

+++ challenge-api/challenge-api/src/main/java/gp/moto/challenge_api/model/Cidade.java
@@ -12,7 +12,7 @@ import org.hibernate.validator.constraints.Length;
 public class Cidade {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idCidade;
+    private Long idCidade;
     @NotEmpty(message = "Valor inválido para nome da cidade")
     @Length(max = 200, message = "Valor inválido para nome da cidade. O valor deve ter no máximo 200 caracteres")
     private String nmCidade;
