+++ challenge-api/challenge-api/src/main/java/gp/moto/challenge_api/model/TipoMoto.java
@@ -13,7 +13,7 @@ public class TipoMoto {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_tipo_moto;
+    private Long id_tipo_moto;
     @NotEmpty(message = "Valor inválido para tipo de moto")
     @Length( max = 40 , message = "Valor inválido para tipo de moto. O valor deve ter no máximo 40 caracteres")
     private String nmTipo;
