+++ challenge-api/challenge-api/src/main/java/gp/moto/challenge_api/model/Uwb.java
@@ -12,7 +12,7 @@ public class Uwb {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idUwb;
+    private Long idUwb;
 
     @OneToOne
     @JoinColumn(name = "id_moto")
